chore: Update buildx-on-pr.yml and buildx.yml workflows (#13)
* chore: Update buildx-on-pr.yml and buildx.yml workflows

Refactor the buildx-on-pr.yml and buildx.yml workflows to improve efficiency and maintainability. Remove the unused IMAGE_NAME environment variable and replace it with a dynamic variable based on the repository name. Also, update the cache configuration for the build-push-action to optimize caching during the build process.

* Refactor buildx-on-pr.yml and buildx.yml workflows for efficiency and maintainability

* chore: Update buildx-on-pr.yml to use dynamic repository name for IMAGE_NAME environment variable

* buildx refactor

* chore: Update buildx.yml workflow to disable pull_request trigger

The pull_request trigger in the buildx.yml workflow was commented out to temporarily disable it. This change was made to prevent unnecessary workflow runs triggered by pull requests. The workflow will now only run on workflow_dispatch and schedule triggers.

---------

Co-authored-by: A <andrei.noga@sva.de>