chore: Add helpers:pinGitHubActionDigests to renovate.json
This commit adds the "helpers:pinGitHubActionDigests" extension to the "renovate.json" file. This extension is added to the existing list of extensions in the file. The purpose of this change is to enable pinning the digests of GitHub Actions used in the project. This will ensure that the project uses specific versions of the GitHub Actions, providing stability and reproducibility.